Move SparkContextFactory to context/ subpackage
