+++ job-server/src/main/resources/application.conf
@@ -21,7 +21,7 @@ spark {
     context-creation-timeout = 15 s
 
     # A zero-arg class implementing spark.jobserver.util.SparkContextFactory
-    context-factory = spark.jobserver.util.DefaultSparkContextFactory
+    context-factory = spark.jobserver.context.DefaultSparkContextFactory
   }
 
   # predefined Spark contexts
