+++ job-server/test/spark.jobserver/LocalContextSupervisorSpec.scala
@@ -20,7 +20,7 @@ object LocalContextSupervisorSpec {
       }
       jobserver.job-result-cache-size = 100
       jobserver.context-creation-timeout = 5 s
-      jobserver.context-factory = spark.jobserver.util.DefaultSparkContextFactory
+      jobserver.context-factory = spark.jobserver.context.DefaultSparkContextFactory
       contexts {
         olap-demo {
           num-cpu-cores = 4
