+++ job-server/test/spark.jobserver/SparkWebUiActorSpec.scala
@@ -34,7 +34,7 @@ object SparkWebUiActorSpec {
       }
       jobserver.job-result-cache-size = 100
       jobserver.context-creation-timeout = 5 s
-      jobserver.context-factory = spark.jobserver.util.DefaultSparkContextFactory
+      jobserver.context-factory = spark.jobserver.context.DefaultSparkContextFactory
       contexts {
         olap-demo {
           num-cpu-cores = 4
